+++ packages/server/graphql/public/types/TeamHealthResultPhase.ts
@@ -1,0 +1,13 @@
+import type {TeamHealthResultPhase as TeamHealthResultPhaseDB} from '../../../postgres/types/NewMeetingPhase'
+import type {TeamHealthResultPhaseResolvers} from '../resolverTypes'
+
+export type TeamHealthResultPhaseSource = TeamHealthResultPhaseDB & {
+  meetingId: string
+  teamId: string
+}
+
+const TeamHealthResultPhase: TeamHealthResultPhaseResolvers = {
+  __isTypeOf: ({phaseType}) => phaseType === 'TEAM_HEALTH_RESULT'
+}
+
+export default TeamHealthResultPhase